php: objects of helpers used instead of functions
