+++ web/documentserver-example/php/config.json
@@ -14,14 +14,14 @@
     ".potx", ".potm", ".pot", ".odp", ".fodp", ".otp", ".rtf", ".mht", ".html", ".htm", ".xml", ".epub", ".fb2"],
 
   "docServTimeout": "120000",
-  "docServSiteUrl": "https://documentserver/",
+  "docServSiteUrl": "http://127.0.0.1/",
 
   "docServConverterUrl": "ConvertService.ashx",
   "docServApiUrl": "web-apps/apps/api/documents/api.js",
   "docServPreloaderUrl": "web-apps/apps/api/documents/cache-scripts.html",
   "docServCommandUrl": "coauthoring/CommandService.ashx",
 
-  "docServJwtSecret": "",
+  "docServJwtSecret": "9GKkwG4oUrswxS0sp1RJ",
   "docServJwtHeader": "Authorization",
   "docServJwtUseForRequest": true,
